✨ refactor : 카드 데이터 및 콘텐츠 타입 정의를 cards.type.ts로 이동 #167
